Add def-package! call for paredit.
When working with lisp code, I wanted to be able to wrap things in
parentheses and do other parenthesis related things.  So far as I know,
this calls for paredit-mode.  Maybe smartparens can do things, but I
couldn't figure that out.  So I set up a def-package! call that
lazily-loads paredit when a lisp buffer appears.  Or at least that's
what I hope I did.