Try adding ein default notebook location.
Compiling emacs lead to an error regarding an ein variable needing to be
set.  Reading ein's README told me to set this variable.  I couldn't get
it to work. This attempt didn't work, and org stuff also was keeping
emacs from totally compiling, so I gave up on the concept.