Update :config default block with upstream changes
From Discord #announcements channel:
As of 9f2dff02, the default smartparens config is gated behind a new
flag (+smartparens) for the :config default module. This config was
enabled by default in the past, so it may seem that delimiters and
parentheses are being autopaired differently (if at all).

If you prefer the old behavior, add +smartparens to your doom! blocks.

Side note: the +evil-commands flag was removed in 4daa9271 (they are now
defined in feature/evil and always enabled). You may safely remove this
flag from :config default.

Relevant commit:
https://github.com/hlissner/doom-emacs/commit/9f2dff02fa8ed3c2162a38195e69aff1085933ea